requirements.txt
# Core RAG and LangChain stack
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.1.49  # ensures latest typing-compatible base
faiss-cpu                # vector store
sentence-transformers>=4.1.0  # local embeddings
llama-cpp-python>=0.2.11      # local LLM engine

# Model-specific dependencies
tiktoken                # for OpenAI embeddings if used
openai                  # fallback or optional use

# Document parsing
pypdf                   # PDF loading
gdown                   # Google Drive file download
python-docx             # optional if you support DOCX in future

# LangChain tools
python-dotenv           # for .env configs
PyYAML                  # for config.yaml parsing

# Telegram bot
python-telegram-bot>=20.0

# Optional for reranking / compression (if needed later)
# cohere    # or bge-reranker, etc.